SleepBot/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.8.1
async-timeout==4.0.1
asyncio==3.4.3
attrs==21.2.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
ciso8601==2.2.0
colorama==0.4.4
colorlog==6.6.0
dnspython==2.1.0
ed25519==1.5
frozenlist==1.2.0
hikari==2.0.0.dev104
hikari-lightbulb==2.1.2
idna==3.3
git+https://github.com/neonjonn/lightbulb-ext-neon.git
multidict==5.2.0
psutil==5.8.0
pure25519==0.0.1
pycares==4.1.2
pycparser==2.21
pymongo==4.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
yarl==1.7.2
uvloop